+++ pkgs/record_use/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Made locations optional to accomodate for dart2js compiler not providing
   source locations for constant instances.
+- Introduce a JSON schema for the json encoding.
 
 ## 0.4.2